Add GCE Ansible authentication style
The new gce_googleapiclient.py dynamic inventory introduces a new way to authenticate with Gooogle Cloud Platform, but it's important to keep the possibility to use the same credentials used by the GCE module from Ansible.

It supports the three different type of authentication supported by Ansible via:
* the secrets.py file in the $PYTHONPATH environment variable
* the gce.ini file
* the GCE_* environment variable

If none of the above authentication options are available, the script uses the Google OAuth lib.