Fix ghost 404 project issue
Even if a project has been shut down, it will still be returned by the billing API for 30 days. The issue is that the project won't be returned by the compute API and produce a 404 when the zones are retrieved. So we should check if the project is active before adding it to the list.